app/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["yansfil <tansfil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.6"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
langchain = "^0.3.1"
langchain-openai = "^0.2.1"
python-dotenv = "^1.0.1"
langserve = {extras = ["all"], version = "^0.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
